+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobiInterstitialAd.m
@@ -120,7 +120,8 @@
                       : GADMAdapterInMobiRequestParametersMediationTypeWaterfall;
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       extras, mediationType,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
   [_interstitialAd setExtras:requestParameters];
 
   if (bidResponseData) {