+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobiUnifiedNativeAd.m
@@ -125,7 +125,8 @@ static CGFloat const DefaultIconScale = 1.0;
                       : GADMAdapterInMobiRequestParametersMediationTypeWaterfall;
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       extras, mediationType,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
   [_native setExtras:requestParameters];
 
   if (bidResponseData) {