+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobiUtils.h
@@ -62,7 +62,7 @@ void GADMAdapterInMobiSetTargetingFromAdConfiguration(
 NSDictionary<NSString *, id> *_Nonnull GADMAdapterInMobiRequestParameters(
     GADInMobiExtras *_Nullable extras,
     GADMAdapterInMobiRequestParametersMediationType _Nonnull mediationType,
-    NSNumber *_Nullable childDirectedTreatment);
+    NSNumber *_Nullable childDirectedTreatment, NSNumber *_Nullable underAgeOfConsent);
 
 /// Logs with GADMediationAdapterInMobi Prefix.
 void GADMAdapterInMobiLog(NSString *_Nonnull format, ...);