+++ adapters/InMobi/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## InMobi iOS Mediation Adapter Changelog
 
+#### Next version
+- Update COPPA implementation to respect the tag for under age consent.
+
 #### [Version 11.1.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/inmobi/InMobiAdapter-11.1.0.0.zip)
 - Verified compatibility with InMobi SDK 11.1.0.